This includes finding any ASM functions that are scattered throughout the codebase and moving them into the asm module, but also reorganizes the asm module so that x86 is a module containing submodules for each native module. This reduces the number of cfg statements needed, and will also make it cleaner to implement assembly for other architectures moving forward.
Closes #1714
