+++ src/asm/x86/mod.rs
@@ -7,8 +7,10 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
-#[cfg(all(
-  feature = "nasm",
-  any(target_arch = "x86", target_arch = "x86_64")
-))]
-pub mod x86;
+pub mod cdef;
+pub mod dist;
+pub mod ec;
+pub mod lrf;
+pub mod mc;
+pub mod predict;
+pub mod transform;
