+++ src/cdef.rs
@@ -19,10 +19,16 @@ use crate::util::{clamp, msb, CastFromPrimitive, Pixel};
 use crate::cpu_features::CpuFeatureLevel;
 use std::cmp;
 
-#[cfg(any(not(target_arch = "x86_64"), not(feature = "nasm")))]
+#[cfg(not(all(
+  feature = "nasm",
+  any(target_arch = "x86", target_arch = "x86_64")
+)))]
 pub(crate) use self::native::*;
-#[cfg(all(target_arch = "x86_64", feature = "nasm"))]
-pub use crate::asm::cdef::*;
+#[cfg(all(
+  feature = "nasm",
+  any(target_arch = "x86", target_arch = "x86_64")
+))]
+pub use crate::asm::x86::cdef::*;
 
 pub struct CdefDirections {
   dir: [[u8; 8]; 8],
