+++ src/lib.rs
@@ -66,8 +66,6 @@ mod deblock;
 mod encoder;
 mod entropymode;
 mod lrf;
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-mod lrf_simd;
 mod mc;
 mod me;
 mod metrics;
